requirements.txt
Flask
python-dotenv
langchain-openai
langchain
langchain-core
